Complete SynapseDispatch module: add DispatchLocationController, MyJobsController, fix migration FK consistency, fix Gantt JS bug
Agent-Logs-Url: https://github.com/Masterleeaus/Worksuite-Saas---Project-Management-System_Laravel/sessions/9138c73e-0e3f-4b0b-a789-4a5dabcb12b0

+++ Modules/SynapseDispatch/Database/Migrations/2024_01_01_000003_create_dispatch_workers_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
             $table->string('code', 64)->unique();
             $table->string('name');
             $table->boolean('is_active')->default(true);
-            $table->unsignedInteger('team_id')->nullable()->index();
+            $table->unsignedBigInteger('team_id')->nullable()->index();
             $table->unsignedBigInteger('location_id')->nullable()->index();
             $table->json('skills')->nullable();
             $table->json('business_hour')->nullable();
